Determination of Non-Random Covariation Residues
This script determines positions in a multi-sequence alignment that have mutations that are not just random but appear together within a given sequence. This can ba a sign of an epistatic relationship. This was some of the first Python coding I did, and I've not looked at it so I'm sure that it's written poorly.